day2
This program stores a list of students with their names and marks in an array of objects. It assumes the first student as the topper, then compares each student’s marks in a loop. Whenever a student with higher marks is found, the topper is updated. Finally, it prints the student who scored the highest marks with a congratulatory message.